% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake NOAA Data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A data frame with the raw NOAA earthquake data.}
}
\value{
A cleaned data frame.
}
\description{
After downloading and reading in the dataset, the overall task for this
module is to write a function named \code{eq_clean_data()} that takes
raw NOAA data frame and returns a clean data frame.
The clean data frame should have the following:
\itemize{
\item A date column created by uniting the year, month, day and converting
it to the \code{Date} class.
\item \code{LATITUDE} and \code{LONGITUDE} columns converted to numeric class.
}
}
\examples{
\dontrun{
eq_clean_data(df)
}
}
\references{
\url{http://www.noaa.gov/}
}
