% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_download_data.R
\name{eq_download_data}
\alias{eq_download_data}
\title{Download Earthquake Data from NOAA}
\usage{
eq_download_data(destfile, data_path)
}
\arguments{
\item{destfile}{A character string with the name where the
downloaded file is saved.}

\item{data_path}{A string with the data path}
}
\value{
None
}
\description{
Download Earthquake Data from NOAA
}
\examples{
\dontrun{
data_path <- "inst/extdata"
dest_file <- "noaa_earthquake.txt"
eq_download_data(destfile = dest_file, data_path = data_path)
}
}
\references{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}

\url{http://www.noaa.gov/}
}
\seealso{
\code{\link[=utils]{download.file}}
}
