% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake NOAA Location Data}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{A data frame with the raw NOAA earthquake data.}
}
\value{
A LOCATION_NAME cleaned data frame.
}
\description{
This function cleans LOCATION_NAME column by stripping out
 the country name (including the colon) and converts names to title case
}
\examples{
\dontrun{
eq_location_clean(df)
}
}
